requirements.txt
langchain-community
langchain
openai
python-dotenv
beautifulsoup4
faiss-cpu
langchain-openai
